requirements.txt
behave
requests
behave-html-pretty-formatter
websocket-client
datetime